Assert noexpandtab and insert start/end times intelligently.
Fix for Issue #1.

Rather than using a playback of keystrokes (e.g. execute "normal
o\<Esc>") to insert start and end times (which is unpredictable
depending upon the user's own settings for autoindent, smartindent,
etc.), calculate the physical indentation of the new start time line,
create the line of text in a variable, and append it to the file.

This effectively sidesteps incompatibilities in the way that indentation
is handled in various Vim environments.

Also, set "noexpandtab" as a local setting explicitly, because all
pattern matching relies upon physical tab characters.