Added changeable text Note to default note
If no existing note is found, a default note '* Note' is added and the word
'Note' is visually selected in select-mode so typing any character will remove
it and one can add whatever note needed.

If a note already exists, simply move
cursor to the last line of the note and start in insert mode to edit the note.