Include digits in snip filename
Just started using this plugin to track dev work that includes tasks with the ticket number. The snip filename was ending up with `---` on the end leaving out the ticket number. This small change seems to resolve the issue but I didn't test much more than trying it out myself.